requirements.txt
absl-py==2.1.0
addict==2.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
attrs==23.2.0
av==11.0.0
cachetools==5.3.2
cffi==1.16.0
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
codecov==2.1.13
coloredlogs==15.0.1
contourpy==1.1.1
coverage==7.4.1
crcmod==1.7
cryptography==42.0.2
cycler==0.12.1
einops==0.7.0
exceptiongroup==1.2.0
facexlib==0.3.0
filterpy==1.4.5
flake8==7.0.0
flatbuffers==23.5.26
fonttools==4.49.0
google-auth==2.27.0
google-auth-oauthlib==1.0.0
grpcio==1.60.1
humanfriendly==10.0
importlib-resources==6.1.1
iniconfig==2.0.0
interrogate==1.5.0
ipykernel
ipywidgets==8.1.2
isort==5.10.1
jmespath==0.10.0
joblib==1.3.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
llvmlite==0.41.1
lmdb==1.4.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mccabe==0.7.0
mdurl==0.1.2
model-index==0.1.11
mpmath==1.3.0
numba==0.58.1
oauthlib==3.2.2
onnxruntime==1.17.0
opencv-python==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.34
ordered-set==4.1.0
oss2==2.17.0
pandas==2.0.3
pillow==10.2.0
pluggy==1.4.0
protobuf==4.25.2
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.20.0
pyflakes==3.2.0
# -e git+https://github.com/datovar4/Image_Quality.git@8915020f355a14c69a2edb850db1dd25e1c3f485#egg=pyiqa_c&subdirectory=pyiqa
pyparsing==3.1.1
pytest==8.0.0
pytest-runner==6.0.1
pytz==2023.4
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.3.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tqdm==4.65.2
transformers==4.38.2
tzdata==2023.4
urllib3==1.26.18
Werkzeug==3.0.1
widgetsnbextension==4.0.10
yapf==0.40.2
h5py
natsort
imageio